[build system]: Fix `USE_SYSTEM_LIBM=1` build
When we started preloading `libm.so`, we neglected to take into account
the name and location of `libm.so` when `USE_SYSTEM_LIBM=1` is used.
This should fix it on all systems except for Ubuntu, which has an `ld`
linker script instead of an actual library for `libm.so`, which has
never worked.